requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydocstyle==6.1.1
pyflakes==3.0.1
pylama==8.4.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
